# docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'Identity Link DB Client API'
  description: 'API documentation for Identity Link DB Clients'
  version: '1.0'
servers:
  -
    url: /clients
    description: 'Clients API base path'
paths:
...
  '/api/v1/client/{id}':
    get:
      tags:
        - Client
      summary: 'Fetch client by ID'
      operationId: 89a0942ff9bbb8a613fa8d7dd83cea98
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 'Client data'
          content:
            application/json:
              schema:
                properties:
                  response: { properties: { client: { $ref: '#/components/schemas/Client' } }, type: object }
                type: object
        '404':
          description: 'Client not found'
    put:
      tags:
        - Client
      summary: 'Update an existing client'
      operationId: a09a41f84e0553f62606f7321aa2630d
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
      responses:
        '200':
          description: 'Client updated'
        '400':
          description: 'Validation error'
        '404':
          description: 'Client not found'
    delete:
      tags:
        - Client
      summary: 'Delete a client'
      operationId: abbd567f9c2a110cc65366d0ea074041
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 'Client deleted'
        '404':
          description: 'Client not found'
  /api/v1/client:
    post:
      tags:
        - Client
      summary: 'Create a new client'
      operationId: c99cfa38ce2e45b3d06d47e842a63e22
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
      responses:
        '201':
          description: 'Client created'
          content:
            application/json:
              schema:
                properties:
                  response: { properties: { client: { $ref: '#/components/schemas/Client' } }, type: object }
                type: object
        '400':
          description: 'Validation error'
  /api/v1/auth:
    post:
      tags:
        - Client
      summary: 'Authenticate client'
      operationId: 6cfe2f02ca10fe11f588bc9198083632
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: 'Authentication successful'
          content:
            application/json:
              schema:
                properties:
                  response: { properties: { client: { $ref: '#/components/schemas/Client' } }, type: object }
                type: object
        '400':
          description: 'Authentication failed'
...
components:
  schemas:
    AuthRequest:
      title: AuthRequest
      description: 'DTO for client authentication request'
      required:
        - id
        - secret
      properties:
        id:
          description: 'Unique identifier of the client'
          type: string
          format: uuid
        secret:
          description: 'Client secret'
          type: string
          maxLength: 200
          minLength: 1
        grantType:
          description: 'OAuth2 grant type'
          type: string
          enum:
            - client_credentials
            - password
            - authorization_code
            - refresh_token
            - implicit
          nullable: true
      type: object
...
    Client:
      title: Client
      description: 'OAuth2 Client entity used for identity link authorization'
      required:
        - audience
      properties:
        id:
          description: 'Client UUID'
          type: string
          format: uuid
        name:
          description: 'Unique client name'
          type: string
          maxLength: 100
        description:
          description: 'Client description'
          type: string
          maxLength: 3000
        audience:
          description: 'HTTP or HTTPS URL identifying the resource server'
          type: string
          format: uri
          maxLength: 3000
        redirectUri:
          description: 'Client redirect URIs'
          type: array
          items:
            type: string
            format: uri
            maxLength: 3000
        grantTypes:
          description: 'Allowed OAuth2 grant types'
          type: array
          items:
            type: string
            enum:
              - client_credentials
              - password
              - authorization_code
              - refresh_token
              - implicit
        groups:
          description: 'Group IDs assigned to the client'
          type: array
          items:
            type: string
        secrets:
          description: 'Secret keys associated with the client'
          type: array
          items:
            type: string
        isPublic:
          description: 'Indicates whether the client is public'
          type: boolean
        isSystem:
          description: 'Indicates whether the client is managed by the system and cannot be updated or deleted through the API'
          type: boolean
          readOnly: true
        consentRequired:
          description: 'Indicates whether the OAuth client requires user consent to be granted via the consent screen during the authorization flow'
          type: boolean
        applicationUrl:
          description: 'Application homepage URL'
          type: string
          format: uri
          nullable: true
        termsOfServiceUrl:
          description: 'Terms of Service URL'
          type: string
          format: uri
          nullable: true
        privacyPolicyUrl:
          description: 'Privacy Policy URL'
          type: string
          format: uri
          nullable: true
        logoUrl:
          description: 'Application logo URL'
          type: string
          format: uri
          nullable: true
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'JWT Bearer token authorization'
      bearerFormat: JWT
      scheme: bearer
security:
  -
    bearerAuth: []
...